store function registry in application state
